cleaning and restoring RootMenuMobile
